Performance improvements and re-enable time scaling.
    - Use far faster LabelAtlas for constant aim updates.
    - Shake field regardless of vibration setting.
    - Implement the new method of time scaling.
    - Improve speed of sky/star rendering by rendering all stars at
      equal distance in one big VBO and updating that VBO's vertex
      positions rather than using a PanningAction.

+++ Classes/WindLayer.m
@@ -98,7 +98,7 @@
     body.scaleX     = wind * windRange * 2 / body.contentSize.width;
     head.rotation   = wind > 0? 180: 0;
     
-    for(uint i = 0; i < [systems count]; ++i) {
+    for(NSUInteger i = 0; i < [systems count]; ++i) {
         ParticleSystem *system = [systems objectAtIndex:i];
         
         if([[affectAngles objectAtIndex:i] boolValue])
